+++ src/Geometry/Rectangle.php
@@ -7,6 +7,7 @@ namespace Intervention\Image\Geometry;
 use ArrayIterator;
 use Intervention\Image\Geometry\Factories\RectangleFactory;
 use Intervention\Image\Interfaces\DrawableFactoryInterface;
+use Intervention\Image\Interfaces\DrawableInterface;
 use Intervention\Image\Interfaces\PointInterface;
 use Intervention\Image\Interfaces\SizeInterface;
 use Intervention\Image\Size;
@@ -44,5 +45,18 @@ class Rectangle extends Size implements SizeInterface
     public function factory(): DrawableFactoryInterface
     {
         return new RectangleFactory($this);
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @see DrawableInterface::adjust()
+     */
+    public function adjust(callable $adjustments): DrawableInterface
+    {
+        $factory = $this->factory();
+        $adjustments($factory);
+
+        return $factory->drawable();
     }
 }