+++ tests/Unit/Geometry/BezierTest.php
@@ -227,6 +227,15 @@ final class BezierTest extends BaseTestCase
         $this->assertInstanceOf(BezierFactory::class, $factory);
     }
 
+    public function testAdjust(): void
+    {
+        $bezier = new Bezier();
+        $this->assertEquals(null, $bezier->backgroundColor());
+        $adjusted = $bezier->adjust(fn(BezierFactory $factory) => $factory->background('f50'));
+        $this->assertEquals(null, $bezier->backgroundColor());
+        $this->assertEquals('f50', $adjusted->backgroundColor());
+    }
+
     public function testClone(): void
     {
         $bezier = new Bezier([