+++ tests/Unit/Geometry/RectangleTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Intervention\Image\Tests\Unit\Geometry;
 
 use Intervention\Image\Alignment;
+use Intervention\Image\Geometry\Factories\RectangleFactory;
 use Intervention\Image\Geometry\Point;
 use Intervention\Image\Geometry\Rectangle;
 use Intervention\Image\Tests\BaseTestCase;
@@ -26,6 +27,22 @@ final class RectangleTest extends BaseTestCase
         $this->assertEquals(-200, $rectangle[3]->y());
         $this->assertEquals(300, $rectangle->width());
         $this->assertEquals(200, $rectangle->height());
+    }
+
+    public function testFactory(): void
+    {
+        $rectangle = new Rectangle(300, 200);
+        $factory = $rectangle->factory();
+        $this->assertInstanceOf(RectangleFactory::class, $factory);
+    }
+
+    public function testAdjust(): void
+    {
+        $rectangle = new Rectangle(300, 200);
+        $this->assertEquals(null, $rectangle->backgroundColor());
+        $adjusted = $rectangle->adjust(fn(RectangleFactory $factory) => $factory->background('f50'));
+        $this->assertEquals(null, $rectangle->backgroundColor());
+        $this->assertEquals('f50', $adjusted->backgroundColor());
     }
 
     public function testSetSize(): void